feat: Migrate from Google Maps to FREE OpenStreetMap
BREAKING CHANGE: Google Maps dependencies replaced with free alternatives

## Summary
- Replaced Google Maps with flutter_map + OpenStreetMap in all Flutter apps
- Added Nominatim for geocoding (free, no API key needed)
- Added OSRM for routing/directions (free, no API key needed)
- Admin dashboard already uses Leaflet with OSM

## Changes by Project

### Flutter Apps (customer-app, delivery-app, restaurant-app)
- Updated pubspec.yaml: Added flutter_map, latlong2, http
- Commented out google_maps_flutter, geocoding, flutter_polyline_points
- Created MapService: Nominatim geocoding, OSRM routing
- Created OSMMapWidget: Reusable OpenStreetMap widget
- Updated map_picker_screen.dart to use OSM
- Updated navigation_screen.dart with real map and routes

### Backend
- Added location.service.ts with Nominatim/OSRM integration
- Updated config with nominatimBaseUrl and osrmBaseUrl
- Added delivery fee calculation using OSRM distances
- Added nearby restaurant/driver search using MongoDB geo queries

### Admin Dashboard
- Created LiveOrdersMap.tsx for real-time order tracking
- Created RestaurantsMap.tsx for restaurant locations
- ZoneMap.tsx already using Leaflet with OSM

## Cost Savings
| Service | Google Maps | OpenStreetMap |
|---------|-------------|---------------|
| 10K map loads | $70 | $0 |
| 10K geocoding | $50 | $0 |
| 10K directions | $50 | $0 |
| TOTAL/month | ~$220 | $0 |

## To run after pulling:
cd customer-app && flutter pub get
cd delivery-app && flutter pub get
cd restaurant-app && flutter pub get

+++ restaurant-app/lib/screens/earnings/earnings_screen.dart
@@ -526,7 +526,7 @@ class _EarningsScreenState extends ConsumerState<EarningsScreen>
                             message:
                                 '${DateFormat('dd/MM').format(earning.date)}\n${_currencyFormat.format(earning.amount)}\n${earning.orders} طلب',
                             child: Container(
-                              height: (heightRatio * 150).clamp(4.0, 150.0),
+                              height: (heightRatio * 150).clamp(4.0, 150.0).toDouble(),
                               decoration: BoxDecoration(
                                 color: AppColors.primary,
                                 borderRadius: const BorderRadius.vertical(
